requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.1.0
    # via
    #   keras
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
blinker==1.9.0
    # via flask
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via flask
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
flask-sqlalchemy==3.1.1
    # via -r requirements.in
flatbuffers==25.1.24
    # via tensorflow
fonttools==4.55.8
    # via matplotlib
gast==0.6.0
    # via tensorflow
google-pasta==0.2.0
    # via tensorflow
grpcio==1.70.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.12.1
    # via
    #   keras
    #   tensorflow
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
keras==3.8.0
    # via tensorflow
kiwisolver==1.4.8
    # via matplotlib
libclang==18.1.1
    # via tensorflow
markdown==3.7
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.10.0
    # via
    #   -r requirements.in
    #   seaborn
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.4.1
    # via
    #   keras
    #   tensorflow
namex==0.0.8
    # via keras
numpy==2.0.2
    # via
    #   -r requirements.in
    #   contourpy
    #   h5py
    #   keras
    #   matplotlib
    #   ml-dtypes
    #   pandas
    #   seaborn
    #   tensorboard
    #   tensorflow
opt-einsum==3.4.0
    # via tensorflow
optree==0.14.0
    # via keras
packaging==24.2
    # via
    #   keras
    #   matplotlib
    #   tensorboard
    #   tensorflow
pandas==2.2.3
    # via
    #   -r requirements.in
    #   seaborn
pillow==11.1.0
    # via matplotlib
protobuf==5.29.3
    # via
    #   tensorboard
    #   tensorflow
pygments==2.19.1
    # via rich
pyparsing==3.2.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
requests==2.32.3
    # via
    #   -r requirements.in
    #   tensorflow
rich==13.9.4
    # via keras
seaborn==0.13.2
    # via -r requirements.in
six==1.17.0
    # via
    #   astunparse
    #   google-pasta
    #   python-dateutil
    #   tensorboard
    #   tensorflow
sqlalchemy==2.0.37
    # via flask-sqlalchemy
tensorboard==2.18.0
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.18.0
    # via -r requirements.in
termcolor==2.5.0
    # via tensorflow
typing-extensions==4.12.2
    # via
    #   optree
    #   sqlalchemy
    #   tensorflow
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
werkzeug==3.1.3
    # via
    #   flask
    #   tensorboard
wheel==0.45.1
    # via astunparse
wrapt==1.17.2
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
